Fix setup.py build on OS X
Previous version failed to use version of standard library with C++11
support, and so choked on attempted include of <type_traits>. Fix is to
explicitly set which standard library to use and ensure min OS X version
is set such that it's willing to use this library.

Note that this is with clang, because OS X g++ is an alias for clang with some
directories in /Developer added to the include path by default.